requirement.txt
asgiref==3.4.1
certifi==2021.10.8
cycler==0.11.0
Django==4.0
django-tables2==2.4.1
djangorestframework==3.13.1
fonttools==4.28.4
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
plotly==5.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
tzdata==2021.5
wincertstore==0.2
